requirements.txt
streamlit
Pillow
torch
transformers
diffusers
numpy
scikit-image
opencv-python
tqdm
streamlit-drawable-canvas 